third_party/blink/web_tests/animations/web-animations/the-effect-value-of-a-keyframe-effect.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Keyframe handling tests</title>
<link rel="help" href="https://drafts.csswg.org/web-animations/#the-effect-value-of-a-keyframe-animation-effect">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/testcommon.js"></script>
<body>
<div id="log"></div>
<div id="target"></div>
<script>
'use strict';

test(function(t) {
  var div = createDiv(t);
  div.style.color = 'rgb(0, 0, 0)';
  var anim = div.animate({ color: [ 'banana-text', 'apple-text'] }, 1000);

  anim.pause();
  anim.currentTime = 0;

  assert_equals(getComputedStyle(div).color, 'rgb(0, 0, 0)');

}, 'Tests keyframes with property values which are unsupported are removed');
</script>
</body>
